Fix: removeEventListener works, function 'onclickType' is isolated
