+++ src/component/LeftBody/Calculator/Calculator.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-console */
 /* eslint-disable no-unused-vars */
 /* eslint-disable no-use-before-define */
 import React, { useState, useEffect } from 'react';
@@ -8,6 +9,7 @@ import CalButtonBox from './CalButtonBox';
 // Function
 import fontVwHalfVh from '../../../responsive/fontVwHalfVh';
 import makeObjState from '../../makeObjState';
+import btnNum from './function/btnNum';
 
 /**
  * 계산기 컴포넌트
@@ -28,10 +30,25 @@ export default function Calculator() {
       refSize: 700,
       vwh: 9,
     };
-    window.removeEventListener('load', (e) => fontVwHalfVh(e, props));
     window.addEventListener('load', (e) => fontVwHalfVh(e, props));
-    window.removeEventListener('resize', (e) => fontVwHalfVh(e, props));
     window.addEventListener('resize', (e) => fontVwHalfVh(e, props));
+
+    let wait;
+    let ing;
+    function down(e) {
+      const value = String(e.key);
+      const keyProps = { states, value, sideClass: 'key' };
+      if (!Number.isNaN(Number(value))) {
+        btnNum(keyProps);
+      }
+    }
+    window.addEventListener('keydown', down);
+
+    return () => {
+      window.removeEventListener('load', (e) => fontVwHalfVh(e, props));
+      window.removeEventListener('resize', (e) => fontVwHalfVh(e, props));
+      window.removeEventListener('keydown', down);
+    };
   }, []);
 
   // Return
